Interrupt running queries on timeout
If a timeout occurs, it leaves the current (long) query running. Subsequent queries
or an attempt to close the connection will block until the query is finished. If new
requests also time out and yet more requests arrive, it is easy to lock the database.

This commit communicates an intterupt (https://sqlite.org/c3ref/interrupt.html) to
the currently running thread to make it abort the running query timely.